Montando el delegate de Detalle de Factura y terminando de implementar los métodos de edición de campos de Edición.

+++ ProyectoFinalCoreContable/MODELO/Factura.h
@@ -49,10 +49,10 @@ Dichas autorizaciones deberán ser objeto de la debida publicidad.
 @property (nonatomic) NSInteger tipoIVA;
 
 /// g) Contraprestación total.
-@property (nonatomic) NSInteger baseImponible;
+@property (nonatomic) float baseImponible;
 
 /// h) En las facturas rectificativas, la referencia a la factura rectificada.
-@property (nonatomic) NSInteger rectificacion;
+@property (nonatomic) float rectificacion;
 
 /// i) En su caso, si se producen las siguientes circunstancias:
 /// - En operaciones exentas referencia a la normativa
